+++ src/models/Settings.php
@@ -1,23 +1,22 @@
 <?php
 /**
- * Document helpers plugin for Craft CMS 3.x
+ * Document helpers plugin for Craft CMS 3.x.
  *
  * Document helpers
  *
- * @link      https://cooltronic.pl
+ * @see      https://cooltronic.pl
+ * @see      https://potacki.com
+ *
  * @copyright Copyright (c) 2021 Paweł Potacki
  */
 
 namespace cooltronicpl\documenthelpers\models;
 
-use cooltronicpl\documenthelpers\DocumentHelper;
-
-use Craft;
 use craft\base\Model;
 
 /**
  * @author    Paweł Potacki
- * @package   DocumentHelpers
+ *
  * @since     1.0.0
  */
 class Settings extends Model
@@ -25,8 +24,7 @@ class Settings extends Model
     // Public Properties
     // =========================================================================
 
-    /**
+    /*
      * @var string
      */
-
 }