+++ src/variables/DocumentHelperVariable.php
@@ -1,26 +1,22 @@
 <?php
 /**
- * Document helpers plugin for Craft CMS 3.x
+ * Document helpers plugin for Craft CMS 3.x.
  *
  * Document helpers
  *
- * @link      https://cooltronic.pl
+ * @see      https://cooltronic.pl
+ * @see      https://potacki.com
+ *
  * @copyright Copyright (c) 2021 Paweł Potacki
  */
 
 namespace cooltronicpl\documenthelpers\variables;
 
-use cooltronicpl\documenthelpers\DocumentHelper;
-
 use Craft;
-use cooltronicpl\documenthelpers\services\DocumentHelperService as DocumentHelperServiceService;
-
-use craft\helpers\Template;
-
 
 /**
  * @author    Paweł Potacki
- * @package   DocumentHelpers
+ *
  * @since     1.0.0
  */
 class DocumentHelperVariable
@@ -29,52 +25,53 @@ class DocumentHelperVariable
     // =========================================================================
 
     /**
-     * @param null $optional
      * @return string
      */
-
     public function pdf($template, $destination, $filename, $variables, $attributes)
     {
-        
-        if(file_exists ( $filename ) && isset($attributes['date'])){
-            if(filemtime($filename) > $attributes['date']){ return $filename;}
-        }        
+        if (file_exists($filename) && isset($attributes['date'])) {
+            if (filemtime($filename) > $attributes['date']) {
+                return $filename;
+            }
+        }
         $vars['entry'] = $variables->getFieldValues();
-        if(isset($variables['title'])){
+        if (isset($attributes['custom'])) {
+            $vars['custom'] = $attributes['custom'];
+        }
+        if (isset($variables['title'])) {
             $vars['title'] = $variables['title'];
         }
         $html = Craft::$app->getView()->renderTemplate($template, $vars);
-        if(isset($attributes['header'])) $html_header = Craft::$app->getView()->renderTemplate($attributes['header']);
-        if(isset($attributes['footer'])) $html_footer = Craft::$app->getView()->renderTemplate($attributes['footer']);
-
-        if(isset($attributes['margin_top'])){
-            $margin_top = $attributes['margin_top'];
+        if (isset($attributes['header'])) {
+            $html_header = Craft::$app->getView()->renderTemplate($attributes['header']);
         }
-        else{
+        if (isset($attributes['footer'])) {
+            $html_footer = Craft::$app->getView()->renderTemplate($attributes['footer']);
+        }
+
+        if (isset($attributes['margin_top'])) {
+            $margin_top = $attributes['margin_top'];
+        } else {
             $margin_top = 30;
         }
-        if(isset($attributes['margin_left'])){
+        if (isset($attributes['margin_left'])) {
             $margin_left = $attributes['margin_left'];
-        }
-        else{
+        } else {
             $margin_left = 15;
         }
-        if(isset($attributes['margin_right'])){
+        if (isset($attributes['margin_right'])) {
             $margin_right = $attributes['margin_right'];
-        }
-        else{
+        } else {
             $margin_right = 15;
         }
-        if(isset($attributes['margin_bottom'])){
+        if (isset($attributes['margin_bottom'])) {
             $margin_bottom = $attributes['margin_bottom'];
-        }
-        else{
+        } else {
             $margin_bottom = 30;
         }
-        if(isset($attributes['mirrorMargins'])){
+        if (isset($attributes['mirrorMargins'])) {
             $mirrorMargins = $attributes['mirrorMargins'];
-        }
-        else{
+        } else {
             $mirrorMargins = 0;
         }
         $pdf = new \Mpdf\Mpdf([
@@ -82,38 +79,46 @@ class DocumentHelperVariable
             'margin_left' => $margin_left,
             'margin_right' => $margin_right,
             'margin_bottom' => $margin_bottom,
-            'mirrorMargins' => $mirrorMargins
+            'mirrorMargins' => $mirrorMargins,
         ]);
 
-        if(isset($attributes['header'])) $pdf_string = $pdf->SetHTMLHeader($html_header);
-        if(isset($attributes['footer'])) $pdf_string = $pdf->SetHTMLFooter($html_footer);
-        if(isset($attributes['pageNumbers'])) $pdf_string = $pdf->setFooter('{PAGENO}');
+        if (isset($attributes['header'])) {
+            $pdf_string = $pdf->SetHTMLHeader($html_header);
+        }
+        if (isset($attributes['footer'])) {
+            $pdf_string = $pdf->SetHTMLFooter($html_footer);
+        }
+        if (isset($attributes['pageNumbers'])) {
+            $pdf_string = $pdf->setFooter('{PAGENO}');
+        }
         $pdf_string = $pdf->WriteHTML($html);
-        if(isset($variables['title'])) $pdf->SetTitle($variables['title']);
+        if (isset($attributes['title'])) {
+            $pdf->SetTitle($attributes['title']);
+        } elseif (isset($variables['title'])) {
+            $pdf->SetTitle($variables['title']);
+        }
 
-        switch ($destination){
-            case "file": $output=\Mpdf\Output\Destination::FILE; break;
-            case "inline": $output=\Mpdf\Output\Destination::INLINE; break;
-            case "download": $output=\Mpdf\Output\Destination::DOWNLOAD; break;
-            case "string": $output=\Mpdf\Output\Destination::STRING_RETURN; break;
-            default: $output=\Mpdf\Output\Destination::FILE; break;
-
-        }    
+        switch ($destination) {
+            case 'file': $output = \Mpdf\Output\Destination::FILE; break;
+            case 'inline': $output = \Mpdf\Output\Destination::INLINE; break;
+            case 'download': $output = \Mpdf\Output\Destination::DOWNLOAD; break;
+            case 'string': $output = \Mpdf\Output\Destination::STRING_RETURN; break;
+            default: $output = \Mpdf\Output\Destination::FILE; break;
+        }
         $return = $pdf->Output($filename, $output);
-        if ($destination == "file"){
+        if ($destination == 'file') {
             return $filename;
         }
-        if ($destination == "download"){
+        if ($destination == 'download') {
             return $filename;
         }
-        if($destination == "inline"){
+        if ($destination == 'inline') {
             return $return;
         }
-        if($destination == "string"){
+        if ($destination == 'string') {
             return $return;
         }
+
         return null;
     }
-
-
 }